split out subscriber notification from publisher
